remove first gps fix at initialization
first gps fix is no longer needed because a session counter is now used to create different filenames.
Also, as the `millis()` timestamp is now the reference for all sensors, actual time can be derived if there has been at least one valid GPS fix during the session.
The objective is to remove the dependency with the GPS for the device to operate.